add memory alloc

+++ source/kernel/include/tools/bitmap.h
@@ -14,13 +14,18 @@
 
 #include "common/types.h"
 
+/**
+ * @brief  位图数据结构
+ * @param bit_count 位图对象管理的内存的分页数量
+ * @param bits 位图数组，记录某一页是否被分配
+ */
 typedef struct  _bitmap_t {
     int bit_count;  //内存分页数量
     uint8_t *bits;  //位图的自己数组
 }bitmap_t;
 
 void bitmap_init(bitmap_t *bitmap, uint8_t *bits, int count, int init_bit);
-int bitmap_get_bit(bitmap_t *bitmap, int index);
+uint8_t bitmap_get_bit(bitmap_t *bitmap, int index);
 void bitmap_set_bit(bitmap_t *bitmap, int index, int count, int bit);
 int bitmap_is_set(bitmap_t *bitmap, int index);
 int bitmap_alloc_nbits(bitmap_t *bitmap, int bit, int count);
